Global header for authenticated users
Updated authenticated axios requests to have a global header axios-auth.js.
In card deck tests I moved test data out into the __mocks__/axios-auth.js
Updated jest.mock to axios-auth instead of plain axios.
Removed un-needed async from should be loading header if isLoaded is false test.
Removed un-needed state call and header config in CardDeck.js.
NPM audit fix and updated misc. dependencies.
